Create centralForm.html
This iteration is enough to present on Monday. It contains all requirements, but it is not a functional form, and it isn't formatted to properly used the whole page yet (you'll have to do some scrolling). Sorry, I didn't realize that this had to be converted into react code in order for it to work in a react app. I assumed that you'd be able to insert it into an existing React page.